Avoid spurious warning for empty 'top nodes'
For address tables with empty 'top' nodes', the address decoder
generator warns about the top node having no width, even if
the width is specified for that node. This is due to the fact that
top nodes are treated specially, and no attributes are considered
for such nodes.

This change fixes the above. Hence no more distracting messages
about missing width specifiers.